requirements.txt
streamlit
pandas
numpy
matplotlib.pyplot
seaborn
pickle
sklearn
plotly.express
wordcloud
